Refactor comparison targets
We don't want to use the word "basis" for the attempt we're comparing against.
It's confusing terminology, the "basis for comparison" is usually time.
Using "target" from now on.

Associate a function with it's description so we can keep the data for the
footer/target generation in one option (speedo-comparison-targets).