compact-mode: Calculate sep width off max col width
Our previous calculation was naive in that it assumed position of line-end and
beginning would result in a fixed length separator string. However, tabulated-list-mode
aligns columns via  a `space` display property. This resulted in a separator
line that would change depending on the current segment. We needed to account for the
length of the segment names in the width calculation.